+++ app/Http/Controllers/ProductController.php
@@ -1,6 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Helpers\RolePermissionHelper;
 use App\Http\Requests\AddProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Models\Product;
@@ -15,8 +17,12 @@ class ProductController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request): View
+    public function index(Request $request)
     {
+
+        if(!RolePermissionHelper::checkPermission("list product")){
+            abort(403);
+        }
       
         $products = Product::with('productCategory')->where(function($query) use ($request){
             if($request->has('search') && !empty($request->search)){
@@ -34,6 +40,10 @@ class ProductController extends Controller
      */
     public function create(): View
     {
+        if(!RolePermissionHelper::checkPermission('create product')){
+            abort(403);
+        }
+
         $product_categories = ProductCategory::all();
         return view('admin.product.create', compact('product_categories'));
     }
@@ -43,6 +53,10 @@ class ProductController extends Controller
      */
     public function store(AddProductRequest $request): RedirectResponse
     {
+        if(!RolePermissionHelper::checkPermission('create_product')){
+            abort(403);
+        }
+
         $imageName = null;  
         if ($request->hasFile('image')) {
             $imageName = time() . '.' . request()->image->getClientOriginalExtension();
@@ -71,6 +85,9 @@ class ProductController extends Controller
      */
     public function edit(Product $product): View
     {
+        if(!RolePermissionHelper::checkPermission('edit product')){
+            abort(403);
+        }
         $products = Product::with('productcategory')->find($product->id);
         $product_categories = ProductCategory::all();
         return view('admin.product.edit', compact('products', 'product_categories'));
@@ -107,6 +124,9 @@ class ProductController extends Controller
      */
     public function destroy(string $id): RedirectResponse
     {
+        if(!RolePermissionHelper::checkPermission('delete product')){
+            abort(403);
+        }
         $product = Product::find($id);
         $product->delete();
